# perf(es/parser): Remove needless `strcmp` ops