--- a/crates/swc_ecma_ast/Cargo.toml
+++ b/crates/swc_ecma_ast/Cargo.toml
@@ -46,6 +46,7 @@ unicode-id = "0.3"
 string_enum = { version = "0.4.1", path = "../string_enum" }
 swc_atoms   = { version = "0.6.0", path = "../swc_atoms" }
 swc_common  = { version = "0.33.3", path = "../swc_common" }
+phf = { version = "0.11.2", features = ["macros"] }
 
 [dev-dependencies]
 serde_json = "1"