--- a/crates/swc_ecma_parser/Cargo.toml
+++ b/crates/swc_ecma_parser/Cargo.toml
@@ -37,6 +37,7 @@ swc_atoms      = { version = "0.6.0", path = "../swc_atoms" }
 swc_common     = { version = "0.33.3", path = "../swc_common" }
 swc_ecma_ast   = { version = "0.110.3", path = "../swc_ecma_ast" }
 swc_ecma_visit = { version = "0.96.3", path = "../swc_ecma_visit", optional = true }
+phf = { version = "0.11.2", features = ["macros"] }
 
 [target.'cfg(not(any(target_arch = "wasm32", target_arch = "arm")))'.dependencies]
 stacker = { version = "0.1.15", optional = true }