--- a/crates/swc_ecma_parser/src/token.rs
+++ b/crates/swc_ecma_parser/src/token.rs
@@ -580,6 +580,7 @@ impl From<&'_ str> for Word {
 }
 
 impl From<&'_ str> for IdentLike {
+    #[inline]
     fn from(s: &str) -> Self {
         s.parse::<KnownIdent>()
             .map(Self::Known)