@Boshen 2023-11-06
You can go further with this approach: https://github.com/web-infra-dev/oxc/blob/a6ef574985ec73e87e3cdbe19a0a0066684626a5/crates/oxc_parser/src/lexer/mod.rs#L1649-L1659 cc @magic-akari 

I managed to cut an enormous amount of CPU cycles.